src/desktop/requirements.txt
paho-mqtt>=2.0.0
psutil>=5.9.0
pyperclip>=1.8.0
Pillow>=10.0.0
pygetwindow>=0.0.9
pycaw>=20230407
comtypes>=1.2.0
python-dotenv>=1.0.0
winotify>=1.1.0
